优化完善店铺管理功能

+++ apis/jd-api/src/main/java/cn/qihangerp/api/jd/mapper/OmsJdGoodsSkuMapper.java
@@ -29,7 +29,7 @@ public interface OmsJdGoodsSkuMapper extends BaseMapper<OmsJdGoodsSku> {
      * @param id 第三方平台设置主键
      * @return 第三方平台设置
      */
-    ShopSetting selectShopSettingById(Long id);
+    ShopSetting selectShopSettingById(Integer id);
 
     void updateShopSessionByShopId(Long shopId,String sessionKey);
 }
