+++ apis/pdd-api/src/main/java/cn/qihangerp/api/pdd/service/OmsPddGoodsService.java
@@ -25,6 +25,6 @@ public interface OmsPddGoodsService extends IService<OmsPddGoods> {
      * @param id 第三方平台设置主键
      * @return 第三方平台设置
      */
-    ShopSetting selectShopSettingById(Long id);
+    ShopSetting selectShopSettingById(Integer id);
     void updateShopSessionByShopId(Long shopId,String sessionKey);
 }
