+++ apis/tao-api/src/main/java/cn/qihangerp/open/tao/mapper/TaoOrderMapper.java
@@ -108,7 +108,7 @@ public interface TaoOrderMapper
      * @param id 第三方平台设置主键
      * @return 第三方平台设置
      */
-    ShopSetting selectShopSettingById(Long id);
+    ShopSetting selectShopSettingById(Integer id);
 
     void updateShopSessionByShopId(Long shopId,String sessionKey);
 }
