+++ apis/tao-api/src/main/java/cn/qihangerp/open/tao/service/ITaoOrderService.java
@@ -73,6 +73,6 @@ public interface ITaoOrderService
      * @param id 第三方平台设置主键
      * @return 第三方平台设置
      */
-    ShopSetting selectShopSettingById(Long id);
+    ShopSetting selectShopSettingById(Integer id);
     void updateShopSessionByShopId(Long shopId,String sessionKey);
 }
